IZUMABL-260 fix coredns not building on inteli7
CoreDNS is not building on inteli7 machine for some reason,
some header files are not found (even though every arm target builds).

Found tips in: https://hackernoon.com/making-a-simple-service-app-for-a-yocto-based-linux-distro

Key fix is: CGO_CFLAGS += "-I${WORKDIR}/recipe-sysroot/usr/include"

Since it's the cgo compilation that fails and this points it to the
right header files.