GHAction: rm build fix for golang
Due to some inventive rules in golang community, all external golang
modules are cloned over as READ ONLY. They cannot be deleted by normal
means. Force them to be writable, so that the rm actually succeeds.

We have a step in the build to make the writable, but that only
takes place if the build succeeds.

https://go.dev/ref/mod

well, actually if you used the

go env -w GOFLAGS=-modcacherw

It would force the modules to be writeable as well.